refactor: refactor action form
